Rewrite part of GetHost function
The previous implentation used TDs indexes in HTML nodes to guess fields
and values. When setting the TTL, the TTL is then returned in additional
TDs and this mess with indexes.

The new implementation parsed all TDs and format the current field to
have a generic map whether some informations are returned or not. Based
on the current, field value are manually cast to the valid type instead
of just returning string as values.

Some functions have been renamed to be more human friendly
(`getNodeText` -> `nodeText` and `getMapValue` -> `try`). Utils
functions have also be moved at the top of the file with a comment.