Wrap HTTP calls to Netmagis and manage errors
This new method `Call` make HTTP calls and handle errors
(`NetmagisError` is returned). Communication/protocol errors are
catched and the returned HTML page is checked for standard error.

Sometimes errors are returned in a more subtil way or you may want to
ensure the action has actually suceeded (by checking for a pattern in
the page). The third parameter of the method is a function definition
that has access to the the HTML body and can do additional parsing. An
error will be return (containing the full raw HTML page) if the
validation fails (if the validation function returns *false*).