Implement retrieving host informations from update form
It reuse `GetHost` name (which previously renamed to `Search`). This
method now parse the update form (GET /mod) for retrieving host
informations. This has the advantage to retrieve an unique identifier
(*idrr*) and to have parameter preformatted for other HTTP POST
requests. This is required to implement the update of a host.