Correct a bug with SMTP rights
When adding or updating a host, to actually remove the permission, the
key *sendsmtp* must be removed from the parameters send to Netmagis.

When retrieving the host informations, the HTML form was also badly
parsed. The value of the <input> is misleading (always *1*) and the
presence of *checked* attribute must be checked.