Manage type of host parameters
Now the `GetHost` method returns the *ttl*, *iddhcpprof*, *idviews*
and *idrr* as int and *sendsmtp* as bool. The input methods (`AddHost`,
`UpdateHost`) still support both specific type for theses parameters or
string.

For an int, *-1* is the "reset" value and is passed as an empty string
to HTTP POST calls. It is required for managing TTL which must be a
positive integer (or a string representing a positive integer) when send
to Netmagis. For the TTL, an empty string remove the value in Netmagis
(but don't know what the fallback value is DNS side).